2.0-beta

+++ dailyWeb/src/main/java/cn/wenzhuo4657/dailyWeb/domain/ItemEdit/model/dto/QueryItemDto.java
@@ -11,13 +11,13 @@ public class QueryItemDto {
     private Long docsId;
 
     @NotNull
-    private int type;
+    private Long type;
 
 
     public QueryItemDto() {
     }
 
-    public QueryItemDto(Long docsId, int type) {
+    public QueryItemDto(Long docsId, Long type) {
         this.docsId = docsId;
         this.type=type;
 
@@ -32,11 +32,11 @@ public class QueryItemDto {
     }
 
 
-    public int getType() {
+    public Long getType() {
         return type;
     }
 
-    public void setType(int type) {
+    public void setType(long type) {
         this.type = type;
     }
 }
